pyproject.toml
[project]
name = "stapply-agent"
version = "0.1.0"
description = "StApply AI Agent and FastAPI server"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Stapply" }]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "python-multipart",
    "python-dotenv",
    "playwright",
    "aiohttp",
    "kernel",
    "browser-use @ git+https://github.com/stapply-ai/browser-use.git@main",
    "anchorbrowser>=0.1.1",
]

[dependency-groups]
dev = []


